made almost no testing progress, but changed some of the file structure
