+++ components/RecCenter.js
@@ -20,16 +20,15 @@ const styles = StyleSheet.create({
   },
 });
 
-class FieldHouse extends Component {
+class RecCenter extends Component {
     render() {
-        const state = this.props.navigation;
-				const navigate = this.props.navigation;
-        const resetAction = NavigationActions.reset({
-            index: 0,
+				const { state, navigate } = this.props.navigation;
+      	const resetAction = NavigationActions.reset({
+          	index: 0,
             actions: [NavigationActions.navigate({ routeName: 'HomeScreen' })],
           });
         return (
-    <Container>
+    	<Container>
             <Header style={{backgroundColor: 'gold'}}>
 	          <Left>
                 <Button transparent dark onPress={() => this.props.navigation.goBack()}>
@@ -37,7 +36,7 @@ class FieldHouse extends Component {
 	            </Button>
 	          </Left>
 	          <Body>
-	          	 <Title>Field House</Title>
+	          	 <Title>Rec Center</Title>
 	          </Body>
 	          <Right>
 	          <Button transparent dark onPress={() => Alert.alert('You Pressed','Forums!')}>
@@ -51,13 +50,13 @@ class FieldHouse extends Component {
 	        <Content>
 	          <ScrollView scrollsToTop={true} ref={(ref) => this.myScroll = ref}>
 	          <View style={styles.container}>
-            	    <Image style={{width: deviceWidth}} resizeMode='cover' source={require('./images/FieldHouse.jpg')}/>
-            	    <Text style={{textAlign: 'center', fontSize: 30}}>{'\nField House'}</Text>
+            	    <Image style={{width: deviceWidth}} resizeMode='cover' source={require('../images/CampusRec.jpg')}/>
+            	    <Text style={{textAlign: 'center', fontSize: 30}}>{'\nCampus Recreation & Wellness Center'}</Text>
             	    <Text style={{textAlign: 'center', fontSize: 20}}>{'\nHours'}</Text>
             	    <Text style={{textAlign: 'center', fontSize: 15}}>{'\nMonday: 5AM - 7PM'}</Text>
             	    <Text style={{textAlign: 'center', fontSize: 15}}>{'Tuesday: 5:30AM - 8PM'}</Text>
-                    <Button style={{alignItems: 'center'}} transparent dark onPress={() => navigate("RecCenter", {screen: "Rec Center"})}>
-                        <Text style={{textAlign: 'center', fontSize: 30, color: 'blue'}}>{'\n\n\n\n\n\n     Take Me To Rec Center'}</Text>
+                    <Button style={{alignItems: 'center'}} transparent dark onPress={() => navigate("FieldHouse", {screen: "Field House"})}>
+                        <Text style={{textAlign: 'center', fontSize: 30, color: 'blue'}}>{'\n\n\n\n\n\n    Take Me To Field House'}</Text>
                     </Button>
                     <Text style={{textAlign: 'center', fontSize: 15}}>{'\n\n\n\n'}</Text>
 	         	</View>
@@ -77,9 +76,9 @@ class FieldHouse extends Component {
     }
 }
 
-FieldHouse.navigationOptions = {
-        title: 'Field House',
-        header: false,
+RecCenter.navigationOptions = {
+    title: `Rec Center`,
+    header: false,
 };
 
-export default FieldHouse
+export default RecCenter;
